Move `qiskit_circuit::util` into own crate
These utility functions were place in the crate originally anticipated
to be the "base" crate.  However, we want ever more quatum-info objects
to be first-class citizens within the circuit IR, which pushes
`quantum-info` to being a lower crate, and consequently implies the
circuit- and quantum-info-independent objects should likely move down
into their own utility crate.
