+++ crates/accelerate/src/sampled_exp_val.rs
@@ -19,10 +19,10 @@ use pyo3::prelude::*;
 use pyo3::wrap_pyfunction;
 
 use crate::pauli_exp_val::fast_sum;
-use qiskit_circuit::util::c64;
 use qiskit_quantum_info::sparse_observable::BitTerm;
 use qiskit_quantum_info::sparse_observable::PySparseObservable;
 use qiskit_quantum_info::sparse_observable::SparseTermView;
+use qiskit_util::complex::c64;
 
 const OPER_TABLE_SIZE: usize = (b'Z' as usize) + 1;
 const fn generate_oper_table() -> [[f64; 2]; OPER_TABLE_SIZE] {
