+++ crates/circuit/src/gate_matrix.rs
@@ -13,10 +13,8 @@
 use num_complex::Complex64;
 use std::f64::consts::FRAC_1_SQRT_2;
 
-use crate::util::{
-    C_M_ONE, C_ONE, C_ZERO, GateArray0Q, GateArray1Q, GateArray2Q, GateArray3Q, GateArray4Q, IM,
-    M_IM, c64,
-};
+use qiskit_util::alias::{GateArray0Q, GateArray1Q, GateArray2Q, GateArray3Q, GateArray4Q};
+use qiskit_util::complex::{C_M_ONE, C_ONE, C_ZERO, IM, M_IM, c64};
 
 pub static ONE_QUBIT_IDENTITY: GateArray1Q = [[C_ONE, C_ZERO], [C_ZERO, C_ONE]];
 pub static TWO_QUBIT_IDENTITY: GateArray2Q = [
