+++ crates/circuit/src/lib.rs
@@ -38,7 +38,6 @@ pub mod parameter;
 pub mod parameter_table;
 pub mod register_data;
 pub mod slice;
-pub mod util;
 pub mod var_stretch_container;
 mod variable_mapper;
 pub mod vf2;
