+++ crates/quantum_info/Cargo.toml
@@ -11,6 +11,7 @@ doctest = false
 
 [dependencies]
 qiskit-circuit.workspace = true
+qiskit-util.workspace = true
 numpy.workspace = true
 num-complex.workspace = true
 num-traits.workspace = true
