+++ crates/util/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "qiskit-util"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[lib]
+name = "qiskit_util"
+doctest = false
+
+[lints]
+workspace = true
+
+[dependencies]
+num-complex.workspace = true
